# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/customers/manage": {
      "get": {
        "operationId": "CustomersManagementController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List customers for the admin dashboard"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      },
      "post": {
        "operationId": "CustomersManagementController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManagedCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create customer manually from the admin dashboard"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      }
    },
    "/customers/manage/abandoned-carts": {
      "get": {
        "operationId": "CustomersManagementController_listAbandonedCarts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List abandoned carts for the admin dashboard"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      }
    },
    "/customers/manage/abandoned-carts/{abandonedCartId}/send-recovery": {
      "post": {
        "operationId": "CustomersManagementController_sendAbandonedCartRecovery",
        "parameters": [
          {
            "name": "abandonedCartId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Send abandoned cart recovery email manually"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      }
    },
    "/customers/manage/{customerId}": {
      "get": {
        "operationId": "CustomersManagementController_getById",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get managed customer details"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      },
      "patch": {
        "operationId": "CustomersManagementController_update",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateManagedCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update managed customer profile"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      }
    },
    "/customers/manage/{customerId}/status": {
      "patch": {
        "operationId": "CustomersManagementController_updateStatus",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateManagedCustomerStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Activate/deactivate managed customer"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      }
    },
...
    "/admin/payment-methods/available": {
      "get": {
        "operationId": "AdminPaymentMethodsController_available",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-payment-methods"
        ]
      }
    },
    "/admin/payment-methods": {
      "get": {
        "operationId": "AdminPaymentMethodsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-payment-methods"
        ]
      }
    },
    "/admin/payment-methods/{paymentMethodCatalogId}/enable": {
      "post": {
        "operationId": "AdminPaymentMethodsController_enable",
        "parameters": [
          {
            "name": "paymentMethodCatalogId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-payment-methods"
        ]
      }
    },
    "/admin/payment-methods/{id}": {
      "patch": {
        "operationId": "AdminPaymentMethodsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStorePaymentMethodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-payment-methods"
        ]
      }
    },
    "/admin/payment-methods/{id}/toggle": {
      "patch": {
        "operationId": "AdminPaymentMethodsController_toggle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleStorePaymentMethodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-payment-methods"
        ]
      }
    },
...
    "/app/payment-receipts/presign": {
      "post": {
        "operationId": "StorefrontPaymentMethodsController_presignReceipt",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorefrontPresignReceiptDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "storefront-payment-methods"
        ]
      }
    },
    "/app/payment-receipts": {
      "post": {
        "operationId": "StorefrontPaymentMethodsController_confirmReceipt",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorefrontConfirmReceiptDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "storefront-payment-methods"
        ]
      }
    },
...
    "/payments/{paymentId}": {
      "get": {
        "operationId": "PaymentsController_getById",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get payment by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommercialErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommercialErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommercialErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Alnjoom Telecom Store API",
    "description": "Backend APIs for Alnjoom Telecom Store",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateManagedCustomerDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateManagedCustomerDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateManagedCustomerStatusDto": {
        "type": "object",
        "properties": {}
      },
...
      "CommercialErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "object"
          },
          "path": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "statusCode",
          "message",
          "path",
          "timestamp",
          "requestId"
        ]
      },
...
      "AllowedTransitionDto": {
        "type": "object",
        "properties": {
          "command": {
            "type": "string"
          },
          "toStatus": {
            "type": "string"
          },
          "requiresReason": {
            "type": "boolean"
          }
        },
        "required": [
          "command",
          "toStatus",
          "requiresReason"
        ]
      },
...
      "UpdateStorePaymentMethodDto": {
        "type": "object",
        "properties": {}
      },
      "ToggleStorePaymentMethodDto": {
        "type": "object",
        "properties": {}
      },
      "StorefrontPresignReceiptDto": {
        "type": "object",
        "properties": {}
      },
      "StorefrontConfirmReceiptDto": {
        "type": "object",
        "properties": {}
      },
...
      "PaymentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "submitted",
              "under_review",
              "approved",
              "rejected",
              "expired",
              "cancelled",
              "partially_refunded",
              "refunded"
            ]
          },
          "statusLabel": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "paidAmount": {
            "type": "string"
          },
          "refundedAmount": {
            "type": "string"
          },
          "refundableAmount": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "submissionVersion": {
            "type": "number"
          },
          "storePaymentMethodId": {
            "type": "string",
            "nullable": true
          },
          "paymentMethodCatalogId": {
            "type": "string",
            "nullable": true
          },
          "paymentMethodCode": {
            "type": "string",
            "nullable": true
          },
          "paymentMethodName": {
            "type": "string",
            "nullable": true
          },
          "accountName": {
            "type": "string",
            "nullable": true
          },
          "accountNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "iban": {
            "type": "string",
            "nullable": true
          },
          "instructionsAr": {
            "type": "string",
            "nullable": true
          },
          "instructionsEn": {
            "type": "string",
            "nullable": true
          },
          "payerReference": {
            "type": "string",
            "nullable": true
          },
          "payerReceiptUrl": {
            "type": "string",
            "nullable": true
          },
          "payerReceiptMediaAssetId": {
            "type": "string",
            "nullable": true
          },
          "payerNote": {
            "type": "string",
            "nullable": true
          },
          "customerSubmittedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "receiptUrl": {
            "type": "string",
            "nullable": true
          },
          "receiptMediaAssetId": {
            "type": "string",
            "nullable": true
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "reviewedBy": {
            "type": "string",
            "nullable": true
          },
          "reviewNote": {
            "type": "string",
            "nullable": true
          },
          "customerUploadedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "allowedTransitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AllowedTransitionDto"
            }
          }
        },
        "required": [
          "id",
          "orderId",
          "method",
          "status",
          "statusLabel",
          "amount",
          "paidAmount",
          "refundedAmount",
          "refundableAmount",
          "currency",
          "version",
          "submissionVersion",
          "storePaymentMethodId",
          "paymentMethodCatalogId",
          "paymentMethodCode",
          "paymentMethodName",
          "accountName",
          "accountNumber",
          "phoneNumber",
          "iban",
          "instructionsAr",
          "instructionsEn",
          "payerReference",
          "payerReceiptUrl",
          "payerReceiptMediaAssetId",
          "payerNote",
          "customerSubmittedAt",
          "receiptUrl",
          "receiptMediaAssetId",
          "reviewedAt",
          "reviewedBy",
          "reviewNote",
          "customerUploadedAt",
          "createdAt",
          "updatedAt",
          "allowedTransitions"
        ]
      },
...
    }
  }
}